# Conflicts:
#	package.json

+++ app/main/event.js
@@ -76,9 +76,10 @@ export default function eventListener(menus) {
   ipcMain.on('create-project', (event, args) => {
     const name = args;
     const folder = `${projectsPath}/${name}`;
-    // const folder = path.join(__dirname, `../documents/projects/${name}`);
     try {
-      // createInitWorkSpace();
+      if (fs.existsSync(folder)) {
+        fse.removeSync(folder);
+      }
       fs.mkdirSync(folder);
       event.returnValue = {
         folder,
@@ -94,8 +95,6 @@ export default function eventListener(menus) {
 
   ipcMain.on('rename-project', (event, args) => {
     const { oldName, newName } = args;
-    // const oldfolder = path.join(__dirname, `../documents/projects/${oldName}`);
-    // const newfolder = path.join(__dirname, `../documents/projects/${newName}`);
     const oldfolder = `${projectsPath}/${oldName}`;
     const newfolder = `${projectsPath}/${newName}`;
     const oldTrashFolder = `${trashPath}/${oldName}`;
@@ -124,21 +123,31 @@ export default function eventListener(menus) {
     const oldPath = `${projectsPath}/${name}`;
     const newPath = `${trashPath}/${name}`;
     try {
-      if (fs.existsSync(newPath)) {
-        const files = fs.readdirSync(oldPath);
-        const fl = files.length;
-        for (let i = 0; i < fl; i++) {
-          const fileName = files[i];
-          fse.moveSync(`${oldPath}/${fileName}`, `${newPath}/${fileName}`, { overwrite: true });
-        }
-        fs.rmdirSync(oldPath);
+      if (!fs.existsSync(oldPath)) { // 不存在对应文件夹
+        event.returnValue = {
+          success: true,
+          code: 1,
+        };
       } else {
-        fse.moveSync(oldPath, newPath, { overwrite: true });
+        if (fs.existsSync(newPath)) {
+          const files = fs.readdirSync(oldPath);
+          const fl = files.length;
+          for (let i = 0; i < fl; i++) {
+            const fileName = files[i];
+            fse.moveSync(`${oldPath}/${fileName}`, `${newPath}/${fileName}`, { overwrite: true });
+          }
+        } else {
+          fse.moveSync(oldPath, newPath, { overwrite: true });
+        }
+        if (fs.existsSync(oldPath)) {
+          fs.rmdirSync(oldPath);
+        }
+        event.returnValue = {
+          success: true,
+          folder: newPath,
+          code: 0,
+        };
       }
-      event.returnValue = {
-        success: true,
-        folder: newPath,
-      };
     } catch (ex) {
       event.returnValue = {
         success: false,
@@ -152,21 +161,11 @@ export default function eventListener(menus) {
     const { name, projectName } = args;
     const file = `${projectsPath}/${projectName}/${name}.md`;
     try {
-      const exists = fs.existsSync(file);
-      if (exists) { // 文件存在
-        event.returnValue = {
-          success: false,
-          error: {
-            errno: -10000,
-          },
-        };
-      } else {
-        fs.writeFileSync(file, '');
-        event.returnValue = {
-          file,
-          success: true,
-        };
-      }
+      fs.writeFileSync(file, '');
+      event.returnValue = {
+        file,
+        success: true,
+      };
     } catch (ex) {
       event.returnValue = {
         success: false,
@@ -178,8 +177,6 @@ export default function eventListener(menus) {
   // 重命名笔记标题
   ipcMain.on('rename-note', (event, args) => {
     const { oldName, newName, projectName } = args;
-    // const oldPath = `${folder}/${oldName}.md`;
-    // const newPath = `${folder}/${newName}.md`;
     const oldPath = `${projectsPath}/${projectName}/${oldName}.md`;
     const newPath = `${projectsPath}/${projectName}/${newName}.md`;
     try {
@@ -257,24 +254,30 @@ export default function eventListener(menus) {
   // 将笔记移动到废纸篓中
   ipcMain.on('move-file-to-trash', (event, args) => {
     const { name, projectName } = args;
-    // const oldPath = `${folder}/${name}.md`;
-    // const newPath = path.join(__dirname, `../documents/trash/${projectName}/${name}.md`);
-    // const newfolder = path.join(__dirname, `../documents/trash/${projectName}`);
     const oldPath = `${projectsPath}/${projectName}/${name}.md`;
     const newPath = `${trashPath}/${projectName}/${name}.md`;
     const newfolder = `${trashPath}/${projectName}`;
     try {
-      if (!fs.existsSync(newfolder)) {
-        fs.mkdirSync(newfolder);
+      if (!fs.existsSync(oldPath)) {
+        event.returnValue = {
+          success: true,
+          folder: newPath,
+          code: 1,
+        };
+      } else {
+        if (!fs.existsSync(newfolder)) {
+          fs.mkdirSync(newfolder);
+        }
+        fse.moveSync(oldPath, newPath, { overwrite: true });
+        if (fs.existsSync(oldPath)) {
+          fs.unlinkSync(oldPath);
+        }
+        event.returnValue = {
+          success: true,
+          folder: newPath,
+          code: 0,
+        };
       }
-      fse.moveSync(oldPath, newPath, { overwrite: true });
-      if (fs.existsSync(oldPath)) {
-        fs.unlinkSync(oldPath);
-      }
-      event.returnValue = {
-        success: true,
-        folder: newPath,
-      };
     } catch (ex) {
       event.returnValue = {
         success: false,
@@ -288,10 +291,18 @@ export default function eventListener(menus) {
     const { projectName, name } = args;
     const filePath = `${trashPath}/${projectName}/${name}.md`;
     try {
-      fs.unlinkSync(filePath);
-      event.returnValue = {
-        success: true,
-      };
+      if (!fs.existsSync(filePath)) {
+        event.returnValue = {
+          success: true,
+          code: 1,
+        };
+      } else {
+        fs.unlinkSync(filePath);
+        event.returnValue = {
+          success: true,
+          code: 0,
+        };
+      }
     } catch (ex) {
       event.returnValue = {
         success: false,
@@ -305,10 +316,17 @@ export default function eventListener(menus) {
     const { name } = args;
     const folder = `${trashPath}/${name}`;
     try {
-      fse.removeSync(folder);
-      event.returnValue = {
-        success: true,
-      };
+      if (!fs.existsSync(folder)) {
+        event.returnValue = {
+          success: true,
+          code: 1,
+        };
+      } else {
+        fse.removeSync(folder);
+        event.returnValue = {
+          success: true,
+        };
+      }
     } catch (ex) {
       event.returnValue = {
         success: false,
